+++ examples/capture_rgb/src/user_app.c
@@ -53,9 +53,6 @@ void user_app(chanend_t c_cam) {
         .config = &config
     };
 
-    // wait a few seconds and ask something
-    delay_seconds_cpp(3);
-
     // set coords and send to ISP
     camera_isp_coordinates_compute(&image);
     camera_isp_start_capture(c_cam, &image);
@@ -78,7 +75,6 @@ void user_app(chanend_t c_cam) {
     config.offset_x = 1.8; 
     config.offset_y = 1.8; 
     camera_isp_coordinates_compute(&image);
-    camera_isp_coordinates_print(&image);
     */
     exit(0);
 }