+++ lib_camera/src/isp/_camera_isp.c
@@ -45,16 +45,22 @@ static struct {
 };
 
 // Sensor width maximum values
-static inline
-unsigned get_max_width(camera_mode_t mode) {
-  switch (mode) {
-  case MODE_RAW:  return MODE_RAW_MAX_SIZE;
-  case MODE_RGB1: return MODE_RGB1_MAX_SIZE;
-  case MODE_RGB2: return MODE_RGB2_MAX_SIZE;
-  case MODE_RGB4: return MODE_RGB4_MAX_SIZE;
-  }
-  return 0; // Invalid mode
-}
+static const unsigned sensor_width_max_values[] = {
+  [MODE_RAW]  = MODE_RAW_MAX_SIZE,
+  [MODE_RGB1] = MODE_RGB1_MAX_SIZE,
+  [MODE_RGB2] = MODE_RGB2_MAX_SIZE,
+  [MODE_RGB4] = MODE_RGB4_MAX_SIZE,
+  [MODE_YUV2] = MODE_YUV2_MAX_SIZE,
+};
+
+// Sensor scale values
+static const unsigned sensor_scale_values[] = {
+  [MODE_RAW]  = 1,
+  [MODE_RGB1] = 1,
+  [MODE_RGB2] = 2,
+  [MODE_RGB4] = 4,
+  [MODE_YUV2] = 2,
+};
 
 // MIPI packet header
 typedef struct {
@@ -109,7 +115,9 @@ void handle_post_process(image_cfg_t* image)
   }
 
 #if (CONFIG_APPLY_AWB)
-  camera_isp_white_balance(image);
+  if (image->config->mode != MODE_YUV2) {
+    camera_isp_white_balance(image);
+  }
 #endif
 
 #if (CONFIG_APPLY_AE)
@@ -164,6 +172,10 @@ void handle_expected_lines(image_cfg_t* image, int8_t* data_in) {
       camera_isp_raw8_to_rgb4(image, data_in, ln);
       break;
     }
+    case MODE_YUV2:{
+      camera_isp_raw8_to_yuv2(image, data_in, ln);
+      break;
+    }
     default:{
       xassert(0 && "mode not supported");
       break;
@@ -175,19 +187,13 @@ void handle_expected_lines(image_cfg_t* image, int8_t* data_in) {
 
 // -------- Image Coordinates --------
 
-inline
-void camera_isp_coordinates_print(image_cfg_t* img_cfg){
-  camera_cfg_t *cfg = img_cfg->config;
-  printf("x1: %d, y1: %d, x2: %d, y2: %d\n", cfg->x1, cfg->y1, cfg->x2, cfg->y2);  
-}
-
 void camera_isp_coordinates_compute(image_cfg_t* img_cfg){
   camera_cfg_t *cfg = img_cfg->config;
 
   // If RAW, scale = 1
   unsigned mode = cfg->mode;
-  unsigned max_size = get_max_width(mode);
-  unsigned scale = (mode == MODE_RAW) ? 1 : (unsigned)(mode);
+  unsigned max_size = sensor_width_max_values[mode];
+  unsigned scale = sensor_scale_values[mode];
 
   // Compute the coordinates of the region of interest
   cfg->x1 = cfg->offset_x * SENSOR_WIDTH;
@@ -215,7 +221,8 @@ void camera_isp_coordinates_compute(image_cfg_t* img_cfg){
 
   // if raw ensure channels are 1, else 3
   unsigned cond_raw = (mode == MODE_RAW && img_cfg->channels == 1);
-  unsigned cond_rgb = (mode != MODE_RAW && img_cfg->channels == 3);
+  unsigned cond_yuv = (mode == MODE_YUV2 && img_cfg->channels == 2);
+  unsigned cond_rgb = ((!cond_raw) && (!cond_yuv) && img_cfg->channels == 3);
 
   // debug info
   debug_printf("Coords: x1:%d, y1:%d, x2:%d, y2:%d\n", cfg->x1, cfg->y1, cfg->x2, cfg->y2);
@@ -223,9 +230,8 @@ void camera_isp_coordinates_compute(image_cfg_t* img_cfg){
   debug_printf("Mode: %d\n", mode);
 
   // ensure everything is logical
-  xassert(cond_raw || cond_rgb && "channels not valid");
+  xassert(cond_raw || cond_rgb || cond_yuv && "channels not valid");
   xassert(cfg->sensor_width <= max_size && "sensor_width");
-  xassert(cfg->sensor_height <= max_size && "sensor_height");
   xassert(cfg->x1 < cfg->x2 && "x1");
   xassert(cfg->y1 < cfg->y2 && "y1");
   xassert(cfg->x2 <= SENSOR_WIDTH && "x2");